+++ deepinv/models/__init__.py
@@ -26,3 +26,4 @@ from .gan import (
     DCGANDiscriminator,
 )
 from .complex import to_complex_denoiser
+from .dynamic import TimeAgnosticNet