+++ deepinv/physics/generator/__init__.py
@@ -12,4 +12,9 @@ from .mri import (
     EquispacedMaskGenerator,
 )
 from .noise import SigmaGenerator
-from .inpainting import BernoulliSplittingMaskGenerator
+from .inpainting import (
+    BernoulliSplittingMaskGenerator,
+    GaussianSplittingMaskGenerator,
+    Artifact2ArtifactSplittingMaskGenerator,
+    Phase2PhaseSplittingMaskGenerator,
+)